requirements.txt
streamlit>=1.28.0
requests>=2.31.0
newspaper3k>=0.2.8
transformers>=4.30.0
torch>=2.0.0
